separated fetch functions from helper functions, crated css folder, edited import paths,edited footer component,app functionality success
